+++ R/predictions.R
@@ -40,6 +40,9 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
   } else if (fun == "gee") {
     # gee-objects -----
     fitfram <- get_predictions_gee(model, expanded_frame, linv, ...)
+  } else if (fun == "clm") {
+    # clm-objects -----
+    fitfram <- get_predictions_clm(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun == "polr") {
     # polr-objects -----
     fitfram <- get_predictions_polr(model, expanded_frame, linv, ...)
@@ -176,6 +179,74 @@ get_predictions_polr <- function(model, fitfram, linv, ...) {
   # No CI
   fitfram$conf.low <- NA
   fitfram$conf.high <- NA
+
+  fitfram
+}
+
+
+# predictions for cumulative link model ----
+
+#' @importFrom sjmisc to_long
+get_predictions_clm <- function(model, fitfram, ci.lvl, linv, ...) {
+  # does user want standard errors?
+  se <- !is.null(ci.lvl) && !is.na(ci.lvl)
+
+  # compute ci, two-ways
+  if (!is.null(ci.lvl) && !is.na(ci.lvl))
+    ci <- 1 - ((1 - ci.lvl) / 2)
+  else
+    ci <- .975
+
+  # prediction, with CI
+  prdat <-
+    stats::predict(
+      model,
+      newdata = fitfram,
+      type = "prob",
+      interval = se,
+      level = ci,
+      ...
+    )
+
+  # convert to data frame.
+  prdat <- as.data.frame(prdat)
+
+  # bind predictions to model frame
+  fitfram <- dplyr::bind_cols(prdat, fitfram)
+
+  # get levels of response
+  lv <- levels(sjstats::model_frame(model)[[sjstats::resp_var(model)]])
+
+  # for proportional ordinal logistic regression (see ordinal::clm),
+  # we have predicted values for each response category. Hence,
+  # gather columns. Since we also have conf. int. for each response
+  # category, we need to gather multiple columns at once
+
+  if (isTRUE(se)) {
+
+    # length of each variable block
+    l <- seq_len(ncol(prdat) / 3)
+    colnames(fitfram)[l] <- lv
+
+    fitfram <- sjmisc::to_long(
+      fitfram,
+      keys = "response.level",
+      values = c("predicted", "conf.low", "conf.high"),
+      l,
+      l + length(l),
+      l + 2 * length(l)
+    )
+
+  } else {
+    key_col <- "response.level"
+    value_col <- "predicted"
+
+    fitfram <- tidyr::gather(fitfram, !! key_col, !! value_col, !! 1:ncol(prdat))
+
+    # No CI
+    fitfram$conf.low <- NA
+    fitfram$conf.high <- NA
+  }
 
   fitfram
 }