+++ R/utils_model_family.R
@@ -30,7 +30,7 @@ get_glm_family <- function(fit) {
   } else {
     # "lrm"-object from pkg "rms" have no family method
     # so we construct a logistic-regression-family-object
-    if (mc %in% c("lrm", "polr", "logistf"))
+    if (mc %in% c("lrm", "polr", "logistf", "clm"))
       faminfo <- stats::binomial(link = "logit")
     else
       # get family info