+++ NEWS.md
@@ -3,6 +3,7 @@
 ## General
 
 * Support for `brmsfit`-models from the **brms**-package.
+* Support for `clm`-models from the **ordinal**-package.
 * Posterior predictive distributions (see argument `ppd`) now compute uncertainty intervals also for non-gaussian models.
 * Use functions from package *sjstats* (link inverse, model frame etc.).
 * If the regression model used weights, `ggpredict()` now computes the weighted mean as typical value for predictors that are held constant.